chore: wip farmer data tidying
